Review Fix #1

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -20,7 +20,7 @@ public interface FilmStorage {
 
     void addLikeByUser(Long filmId, Long userId);
 
-    Collection<Film> getTopFilms(Integer count, Integer genreId, Integer year);
+    Collection<Film> getTopFilms(Integer count, Long genreId, Integer year);
 
     Collection<Film> getCommonFilms(Long userId, Long friendId);
